requirements.txt
streamlit
python-dotenv
crewai
langchain
langchain_community
langchain_google_genai
faiss-cpu
